[AUA] use CheckboxTree for Tool Window display
The checkboxes are (intended to be) only enabled if checking or
unchecking them at that point makes sense from the point of view
of the Upgrade Assistant.

This now provides a superset of the functionality of the flag-off
Upgrade Assistant, though the UI is rudimentary and the various UX flows
are not polished.

Bug: 169647448
Test: not yet
Change-Id: Id4f33932096add9097fa32ca3d8c5ec96c3ab5af