Add a 30ms timeout to executeCallbacksAndRequestRender
When the callback passed to executeCallbacksAndRequestRender takes more than
30ms to run, cancel the execution and return early to prevent freezing the
UI. These callbacks shouldn't take that long to run but they could be
waiting for Layoutlib Render Thread to execute other heavyweight operations,
such as inflating the layout. Callers of executeCallbacksAndRequestRender
now need to check the return value to verify if the callback was executed
successfully and react accordingly otherwise.

For now, the Animation Preview ignore failed callback executions, which
should be fine for most cases. However, in the future we will need to handle
the first call to "update animation states" in case it takes too long,
otherwise we won't be able to track animations until the animation preview
is restarted.

Fixes: 173210615
Test: ComposePreviewAnimationManager should pass and manually verified the
      UI does not freeze if Animation Preview operations take too long.
Change-Id: I06ad9040f11eac3ea8bffabe4e66c0dec58dede8

GitOrigin-RevId: 682dee3f2d1e09c7a8d3fb2ddbcfab67a6027a6c