Add Android Studio linux start-up test
This change adds a test that launches the production
Android Studio linux binaries. It launches
the IDE with an empty configuration and monitors
idea.log until the PluginManager loads all plugins.

Then it asserts that all the loaded plugins are the
expected ones.

Run the test hundreds of times locally and on RBE
to determine the current timeouts.

Test: this is it
Bug: NA
Change-Id: I39291dc11e68de7caa235feb0bd7ad9cabbef3db