[AUA] Implementation/Api configurations available earlier
The Implementation/Api configurations (and related ones for non-main
sourcesets) were introduced in Gradle 3.4, and were available since
then.  The Upgrade Assistant (and versions of Studio it runs in)
doesn't in fact support anything before Android Gradle Plugin 3.1, but
even Android Gradle Plugin 3.0 required Gradle version 4.4 or later,
so that is a safe lower bound for computing the necessity of replacing
deprecated configurations in the Upgrade Assistant.

Bug: n/a
Test: existing tests adjusted
Change-Id: I2ccc2051914fa7ae0b81701ff03722c6a5bcb1a3

GitOrigin-RevId: 91a53ed15fcb743e7c2ee2d764dc01b11bf32cda