Support resolving artifacts against google3
This change introduces a ModuleSystemArtifactResolver that can resolve
artifacts against the underlying build system. For blaze projects,
libraries like androidx workmanager are resolved directly against the
user's google3 workspace directory.

This new resolver is only used in blaze projects, and only when the
desired artifact can't be resolved on GMaven.

Test: //tools/adt/idea/appinspection/...
Bug: 173713873
Change-Id: If9e1e8ce5635d5244f56164527aae521e197c8b4