[auto-import] Can provide multiple fixes
When given one class name, we may have multiple
choices in terms of adding associated dependecies,
based on the index data.

Now, the behavior is like this. If there's only
one choice, we just show the explicit fix, e.g:
"Add dependency on <maven coordinate>". Else,
we show "Add library dependency", and a popup
for users to choose from.

We are able to provide fixes for either short
class names or fully qualified class names.

Also, for testability, we covert GMavenIndexRepository
into an application service.

Bug: 173547479
Test: Updated existing tests.
Change-Id: I25cecf69b3060184e32f8790f7dc008ede4d72b7

GitOrigin-RevId: 7b2c405939f80a2408d0404ec50fb51e5d0be3ce