[AUA] add argument to dialog constructor
The argument will be used to selectively show a warning message to the
user in the event that there are significant uncommitted changes in the
project.  (This change simply introduces the argument, and defaults it
to false everywhere; a subsequent change will initialize and use it
appropriately).

Bug: 173587367
Test: existing (no change in functionality)
Change-Id: Ic323812c2edc9c437b6a2c3e5319ab8799e6a4d6
(cherry picked from commit 95a39a6625e8b42585137f703201237cf9bae3e1)