[Dsl] parse version/apply infix operators
Extend the Groovy and KotlinScript parsers to be able to parse the
specific case of `version` and `apply` infix operators, and to maintain
the properties associated with those in the GradleDslInfixExpression
structure (which is somewhat analogous to a ExpressionList or
ExpressionMap, in that it is a container for properties).

There is an unfortunate mismatch between the Dsl structures and the
underlying Psi in this case: the Dsl in general maintains a
correspondence between Dsl properties and a single Psi element, but in
this case in neither build language is there a single Psi element whose
contents are the infix operator and its right-hand operand; instead, at
the relevant nodes, we have a left-leaning binary tree (in Groovy) or a
ternary tree (in KotlinScript).  In practice this means that deleting
such a property in isolation will probably do the wrong thing.

However, this is enough that the hooks in the PluginModel now function
at a basic level, so we can read version and apply properties.

Bug: 165576187
Test: strengthen one test, more in subsequent CLs
Change-Id: I71bfbf340c641bfb482f52b51ebf4a3419bebc38

GitOrigin-RevId: 38c955e852cd8fbbac720f048c635482d5058490