Disable AC and ACC buttons when project needs syncing
Currently, if the project needs sync'ing and the user attempts to
Apply Changes, it will result in an error since no program runners
are available to execute the configuration.

This patch adds a fix by checking if any program runners are
available before enabling the AC/ACC buttons (if they should be
enabled otherwise).

Fixes: 174124219
Fixes: 165924126
Test: manually verified by forcing a sync needed and then syncing
Change-Id: Icc2cd27c8c9fdece03c6c405c19164d5fc871c1c

GitOrigin-RevId: 0ac6ca6f43031d37a1c647fe677d69ac18bad475