Make sure EDT is used before trying to change JDK
... when using STUDIO_GRADLE_JDK.

There is a deadlock caused by not preserving the order of EDT and a lock
that protects settings related to STUDIO_GRADLE_JDK. When
STUDIO_GRADLE_JDK is defined and is valid, IdeSdks creates a JDK (using
Jdks#createJdk) with the path this environment variable points, storing
the result in myEnvVariableJdkSdk. In order to prevent multiple instances
to be created if two or more threads are called, any read and write to
settings related to STUDIO_GRADLE_JDK are protected by myEnvVariableLock.
When Jdks#createJdk is called, it needs to use EDT to create the Sdk
since it is a write action.

If the EDT tries to use settings related to the environment variable
while another thread has started the initialization, then the deadlock
happens because EDT is waiting for myEnvVariableLock while the other
thread holds it and is waiting for EDT to be available to create the
JDK.

This change prevents the deadlock by first releasing myEnvVariableLock
before trying to create the JDK using the EDT and making EDT to get the
lock instead. If the initialization is called by EDT then this happens:

EDT -> lock aquired -> lock released -> lock aquired -> JDK created ->
lock released

If the initialization is stared by another thread, then this happens

lock aquired -> lock release -> EDT -> lock aquired -> JDK created ->
lock released

Now in both cases the order is the same and the lock is not held before
using EDT. EDT is used only when it is safe to do so (outside of read
actions or if the read action has the intend to write lock).

This change also moves all settings related to the environment variable
to a new private class to simplify the code.

Bug: 174675513
Test: IdeSdksTest and IdeSdksAndroidTest
Change-Id: I50a8f43a05883cae8d6ba60b692c508be0abfed6
(cherry picked from commit 261a3281fd4cab32e2507ca39b86d7976030cbc9)