Refactor caching out of artifact resolver code
The main change is the decoupling of caching from resolver code, because
other classes (such as dexing) will need to use it too. In addition,
this got rid of some boiler plate code in the resolvers.

Now, resolvers focus on resolving, and that's it. Cache handling has
been pulled up to the new InspectorArtifactService. This new application
service is designed to hide the complexity in resolving artifacts, from
deciding which resolver to use to keeping track of cached artifacts.

Overall, the goal of this change is to make the code more modular and
easier to work with.

Other notable changes:
- AppInspectorJarPaths renamed to AppInspectorArtifactPaths because it
caches AARs and dexes in addition to jars.
- ArtifactResolverFactory is no longer a service.

Test: Added new InspectorArtifactServiceTest.
Bug: 165319681
Change-Id: I54a04cd54445d38e8de56bcf907bf2c4f6a22587