Apply the recommended Kotlin code style settings
I upgraded to IDEA 2020.3 (which presumably also upgraded my Kotlin
plugin). After which I kept getting a nag to use the recommended Kotlin
code settings. This change is what "Apply the code style" does.

I went to a meaty Kotlin file, reformatted the whole thing, and it
didn't change anything. So maybe this is a no-op?

Bug: Not applicable
Test: Not applicable
Change-Id: I0e6ed9204376df30a52aa4d473b94c95c1d1d9d1

GitOrigin-RevId: 01def62d96f59df4eb0ff756fecb593cf2cdb8fb