Support for SDK Manager UI as a standalone tool
This change contains:
1. Hitting "Finish" after installing the SDK in a brand new folder
should update the SDK Manager contents. This was a bug that also existed
in Android Studio, and can be reproduced by (1) Skip SDK Wizard, (2) Open a
project, (3) hit SDK Manager, (4) hit Edit, (5) Install SDK, (6) Hit
Finish and then the packages just installed by the SDK Wizard are not
checked in the list.

2. Standalone SDK Manager UI should not allow changing the location of
the SDK directory in the SDK wizard. The SDK path is passed in as a
parameter to the standalone tool (to be passed in by AGDE) and is fixed.

Bug: 138408848
Test: existing and manual
Change-Id: Ia4f0b73578e662357d6df8ed0ac82a3238750d3f