Replace InspectorProcessManager with App Inspection discovery
This commit removes the process notification / management code in layout
inspector and instead leans entirely on the mechanism provided by the
app inspection module.

This was not a smooth cut and required a fair amount of code surgery,
as the process logic was fairly embedded in the InspectorClient
interface and pulling it out caused a bunch of ripples.

That said, this change allowed us to delete a lot of code and remove all
of the retry logic, as that is automatically handled by the process
notification system we're now building on.

There was also some related cleanup where various fields are initialized
early and passed in, instead of being created or casted at runtime. The
biggest example of this is ADB, which is now created once and can be
used immediately throughout the rest of the layout inspector pipeline
code. TreeLoader code received similar treatment, making it impossible
to create their subclasses with invalid states.

Despite all this, a majority of this CL is actually test code. I have
kept the intent behind all tests exactly the same, but some setup
related to those tests (such as connecting processes) have changed.

Additionally, some effort was made to split apart the previous
LayoutInspectorTransportRule into two separate pieces, a common rule
and a transport-specific rule. In this way, we could guarantee that
the legacy tests don't accidentally make use of transport side
effects, and we will be able to carry this idea forward to the app
inspection pipeline which we'll be adding later.

Fixes: 150618894
Fixes: 176167310
Bug: 174765561
Test: Tests updated in response to process management changes
Change-Id: I6d859dd5fa05a5d9b38ed9a98018bf9705a40b37

GitOrigin-RevId: 38cbff6378fea0923d16a655780697b1ac0134c0