Add DeviceDescriptor to ProcessDescriptor
... and also add apiLevel into the device information, removing the need
to force cast the ProcessDescriptor to a TransportProcessDescriptor in
some places.

Although app inspection doesn't necessarily need this refactoring, it
doesn't hurt, and it will be useful later when we migrate layout
inspector over to using app inspection. The division of device and
process will make it easier to add more fields later, if necessary,
e.g. for device analytics purposes in layout inspector.

Bug: N/A
Test: Covered by existing
Change-Id: I0e392b993c4df526a088ac3df1faad486a87b8b1

GitOrigin-RevId: 0a9786ebb9f3f045e4c4ab3f007318bd93ecdba2