Expand tree root node when empty child is added
When a database without a schema is added, it is not possible
to expand that node (it's an empty node).
Since we hide the root of the tree, when all the children of the
root are empty nodes, they are not visible, becase the root
is collapsed and none of the children can be expanded.

This change expands the root node when a database without a schema
is added.

Test: DatabaseInspectorViewImplTest.kt
Bug: 174990761
Change-Id: Iaa3e43cfe2021581ed5feb4b1235f4f2874547ef

GitOrigin-RevId: faa4848cc29e632cfb38e40c6964d02174e79fb6