Update process discovery to be more general
In app inspection, we use process discovery only for O+ processes, but
in layout inspection, we'd like to lean on the same code path for
additional (earlier) processes.

While doing this refactor, we move some code out of ide to api, since
there's actually no IDE specific logic in it, and it moves it closer to
the other process-related classes (which already live in api).

Bug: N/A
Test: Covered by existing
Change-Id: I60db250b16a98dad4b31b631e97876f8c91ad005

GitOrigin-RevId: 758e62d89f2f8d4d41020fbe6879d66ab5efc371