[Firebase] Fix UI freeze in ModuleTemplateDataBuilderKt
... calling ModuleTemplateDataBuilderKt.getExistingModuleTemplateDataBuilder
makes a call to KeystoreUtils.getSha1DebugKeystoreSilently()
that is very slow, freezing the UI.
The debug key is only used by the Maps Activity and
is not needed in the context of "Existing Module Template Data".
When we add a map activity to an existing module,
RenderTemplateModel already makes a call to
getSha1DebugKeystoreSilently(), so we can remove this call
from getExistingModuleTemplateDataBuilder().

Fixes: 174874816
Test: TemplateTests
Change-Id: I2772bf42281ed5c34022ff9ab3f0cbfb4f1925a3

GitOrigin-RevId: 5377959dd81e8432d3b49d6844c609dafa785324