Fix tests after last IntelliJ 2020.3 merge
The behaviour of mockito has changed slightly and now tests break in
certain cases where we were not mocking the correct calls.
By default, if not mocked, mockito is not calling the real method
unless instructed to do so. This was working before but broke in
this last 2020.3 merge. The commit fixes a number of tests broken
because of that.

Bug: 175036532
Test: The updated tests pass
Change-Id: I326ae6819adfc370b4a929178cc61ced7b9d8094

GitOrigin-RevId: 65fccb6000f0a3dd47c59f10b296ba09f42db279