Remove obsolete constraint layout detector version check
Lint had some version comparisons to make sure users were using at
least 1.0.2.  This was important in early days when there were serious
problems in the layout editor since the constraint layout in the IDE
was not loaded from the project.  That's no longer issue, and
ConstraintLayout does not need special handling; the generalized "new
version available" check for all libraries in GradleDetector is
enough.

Test: Existing
Bug: None
Change-Id: I02d6a132cd8c73cc10196ea1ff358716d7720fa0