Remove explicit additiona/removal in SceneViewPanel
When adding/removing models in the DesignSurface, it would generate an
immediate removal of the SceneViewPeerPanel in the SceneViewPanel. This,
in turn, would generate a repaint and cause the panels to move around
when multiple add/remove model happened in sequence.

With this change, the SceneViewPeerPanels are generated lazily on layout
of the SceneViewPanel. This avoids that a change causing everything to
move around and also avoids needing to have to keep a parallel track of
updates to the panels.

Bug: N/A
Test: Existing tests pass, visually verified
Change-Id: If00d95ae70c4ed7af275fe8aea7a8184395addc9

GitOrigin-RevId: 4856b9b0d43dff739dbfae603b373b72adf52ee5