Change InspectorClient interface to be more pipeline-agnostic
Before, the InspectorClient interface was coupled somewhat tightly with
transport pipeline concepts (registering against a transport event,
executing transport commands), but with the upcoming app inspection
migration, we need to make it a bit higher level.

This commit makes this happen - existing clients still work, and we're
now ready to add a third implementation that is built on top of the app
inspection framework.

Bug: 174765561
Test: Covered by existing
Change-Id: I4c84e027b2b19107f8798875b8afe1610346d7f3

GitOrigin-RevId: 908eb783f31e7fb2f61fe331a5d7996bf7f61b7d