Move the `isCapturingModeOn` global out of the `transport` package
`isCapturingModeOn` is a setting that transcends the transport pipeline,
which is where the property currently lives, because a user can still
influence it even when disconnected, for example. And the future app
inspection pipeline client will use it as well.

Additionally, we introduce a class for it to live inside, so it's no
longer a floating public global.

With this change, there are no more accesses to internal pipeline
packages (e.g. pipeline.transport) outside of the pipeline package.

Bug: 174765561
Test: Covered by existing
Change-Id: I0e34061fca855ef81a4e09f4d2c0bc30f4d417b2

GitOrigin-RevId: 669feaa337041467481bd53b8f17074920780d0e