[Dsl] fix NPE in plugin dsl parser
In KotlinScript Gradle build files, the `kotlin("foo")` shorthand can be
used to specify plugins in the plugin blocks.  This was already handled
by the model for the non-infix syntax; the infix syntax model was not
defensive enough about the potential absence of an `id("foo")` call.

This change defends against the null pointer exception, but does not
correctly model the kotlin call.

Bug: 174793640
Test: enough that the NPE is no longer present
Change-Id: Iaf234ae8b1bd3df0a492e0ce9c12a19a041eae81

GitOrigin-RevId: de09740f4498d281053943c8e04ad17af47bf9df