[Dsl,.kts] handle kotlin plugin function with infix
Adjust the special-case in extractValue() for KotlinScript build files
to handle not only when the direct parent of a kotlin(...) call is the
plugins block, but also when the parent is an infix expression whose
parent is a plugins block.

When building the PluginsModel, if there is no `id` element, check for a
`kotlin` element (strictly speaking we should only do this for
KotlinScript and not Groovy, but slightly-permissive parsing is not a
problem in practice).

Modify a test to check that we can parse kotlin(...) expressions in
plugins block losslessly.

Bug: 174793640
Test: existing, added
Change-Id: Ieb728abcdbc7421cd4cd46fb4073a6fe3c4e38c9

GitOrigin-RevId: 992a35b80b8acd1b83319782a61f20874012a9c2