[AUA] upgrade a plugin only if we can truly parse its version
It turns out that GradleVersion.tryParse() is very permissive about the
contents of its version strings.  "kotlin_version" or "Versions.kotlin"
are not parse errors (for which `tryParse()` would return null) but are
treated as valid version strings with major version 0 (and minor version
0 in the case of "Versions.kotlin").  This is an issue if the user has
defined a variable being used to keep Kotlin versions in sync in a place
or a manner that the Gradle Dsl parser cannot find or resolve.

When the Gradle Dsl parser cannot resolve such a variable, then, the
presumed version of the plugin will be 0.  Catch that case specifically
before investigating further.

Fixes: 174793640
Test: added
Change-Id: I7963545434b4c1ef6b0337f747d69b6ba8a03a67
(cherry picked from commit 0ff5eae8b1b793a187fb603a3099c7db7f948676)