Add device disconnection support in app inspection
A process being removed triggered the "process disconnected" listener,
and now, so does a device being disconnected (e.g. emulator shut down,
USB cable yanked)

This fixes behavior where, if you're using app inspection and the device
becomes unavailable, the process is updated as dead, whereas before it
gave you the illusion it was still running and you could interact with
it.

Bug: N/A
Test: Added
Change-Id: Ifc1627471e61d591c8963e5ebfb3691c81d643fe

GitOrigin-RevId: 9f4ae1557ecd702880643a9ec4494c76f028301b