Use kotlin visitor to locate lambdas
When locating lambdas in library code we were using stubbed source
files for locating lambdas. The line numbers did not match and we
failed to locate the lambda.

The source file name is now specified from the agent which allows us
to use the file index to find the source file.

Fixes: 175397027
Test: Unit tests updated and should pass
Change-Id: Id9609567f7418bd929363e34957d96589fa7be42