Skip nested lambda top elements
The synthetic class name of a lambda is used to differentiate
lambdas found on the same line.

The synthetic name may contain the closest parent that is either:
- a function
- a property
- a class

We will call that a "top element" of the possible lambdas.
When choosing between multiple lambdas for a single line, we start
searching from the closest top element and we should not recurse
into other top elements, which is what we did before this change.

Fixes: 176117255
Test: Unit test updated to cover this problem.
Change-Id: Ic5ced9057f251497b635c47d3c74c8e8663abfc8

GitOrigin-RevId: 80eccbd58ad036512f4cab5724cf2a4978f3d75e