Fix test issues finding ActivityMainBinding after 203.4449.2 update
The issue is that data binding / view binding code expects a
ResolveScopeEnlarger to be called at some point (BindingScopeEnlarger
for Java files, KotlinBindingScopeEnlarger for Kotlin files), which in
turn allows us to find the ActivityMainBinding class, as it will be
confirmed in scope.

In our tests, we currently fetch Kotlin files via a JavaPsiFacade, which
returns a Java view of the Kotlin class. This *used* to trigger
resolve scope enlargers, but it seems it no longer does.

I'm not certain about what upstream change caused this issue, but it
doesn't seem to affect production. To avoid the issue in our tests,
we now resolve Kotlin classes directly via Kotlin plugin APIs.

Fixes: 176092555
Test: @Ignore code re-enabled
Change-Id: I5dde503b268ee475ebe6270785d20905eeacfb04

GitOrigin-RevId: 5d9d4ddcf453016b4ec863bfcc6f29591a63b433