Optimize text fit calculation in node renderer
In the capture node renderer we calculate the pixel width of node
names to determine how we shorten it (e.g. a.c.MyClass.MyMethod).
Turns out FontMetrics#stringWidth is actually quite expensive and
has time complexity linear to the string length, which translates
to huge overhead when many nodes are being displayed but most of
them are too small in pixels to display their names.

This fix short-circuits the width calculation when the rectangle
width is sufficiently small (<= 1 px). This has resuled in 95%
time reduction (4s -> 200ms) in the worst case scenario.

Bug: 157656070
Test: added
Change-Id: I42d88ab8aae06c9462ff3d2b6967db14039aee2b

GitOrigin-RevId: 4b61dceea4f700d1ad5a49e22de935947b25a53d