Fix getTargetSelectedWithComboBox
Return the persisted target instead of the device's default. At the two
points the persisted target matches one of the known devices but it
could still be different than the device's default. A virtual device's
default target is a QuickBootTarget and returning that is wrong if the
persisted target is a ColdBootTarget.

Add a matches method to Target. Matching a BootWithSnapshotTarget with a
device needs to look at the snapshots as well, not just the device key.

Bug: 120497509
Test: BootWithSnapshotTargetTest and DevicesSelectedServiceTest
Change-Id: I284619567be39b221832d97481180c435e66fb52

GitOrigin-RevId: 3abb08b3e4d750e7ba1b87a30a01906e015e36e4