Ensure dialogs are drawn when shown
When we get an update to an existing window in the layout inspector, we
draw it.
When we get a new window, though, we zoom to fit, which then calls the
zoom change listener, which draws the window.
If we get a new dialog, we're already zoomed to fit, and so the zoom
listener wasn't being fired.
After this change, we draw explicitly after zooming to fit, if the zoom
wasn't actually modified.

Bug: 174331954
Test: Added a couple tests
Change-Id: I0536886b0735cd614f422eafa74ec3f06e5f19b7