[sync] Don't use cached models if we should force an upgrade
We try to elide a sync on project startup if possible, which is good,
but causes confusion if the user was expecting a prompt to force an
upgrade (because they are using a new preview release, for example).
Add a validation step for each model before attaching, and implement a
check for the AndroidModuleModel that it is produced from a version that
we would not otherwise force an upgrade from.

Bug: 153680188
Bug: 155109633
Bug: 160258052
Bug: 174787527
Test: manual
Change-Id: Id5b3903d80e6a23ea4e9dda2d559a3b09d2952b4

GitOrigin-RevId: f287d106322db492e2fa3ce0ebc7685551b2dbaf