[sync] Recommend upgrade from cached load as if after sync
If we would have recommended an upgrade on performing a full sync, make
sure to recommend after attaching cached models.  (The recommender is
idempotent, so it is OK to call it for each cached model).

There's probably a wider pattern, here and with the forced upgrade, of
things that should be done before / after sync being simulated at the
right time in loading models from cache.

Bug: n/a
Test: manual
Change-Id: I567a67df936932261dfda7f1b5a8c89224be4bbf

GitOrigin-RevId: e7a9e1919e64f53d7c83cb7443879cfe929c72d4