Avoid loading classes in ClassConverter
While executing the class transformations, the ClassWriter that outputs
the result might ask the class loader to load additional classes to do
calculations needed to generate the final method stack frames.
This causes, in some cases, that the class loader is asked to load the
class that is currently in the process of transforming and fails to do
so.
This CL removes the dependency of the ClassWriter into the ClassLoader
by allowing to do the type calculation without using the class loading
mechanism.
The new system just reads the type information from the .class file
without relying on the Class objects.

Fixes: 176938212
Test: Additional tests for the new PseudoClass infrastructure
Change-Id: I72886499737acc29e3b3a88608a192f842399179