Remove the call to Device::hasKeyContainedBy
Remove the call to Device::hasKeyContainedBy from
DeviceAndSnapshotComboBoxExecutionTarget::filteredStream. One of the
reasons that method was added was to work around the fact that some of
the persisted keys were nonprefixed.

That won't be the case with the switch to PersistentStateComponent. So
replace the call with myKeys.contains(device.getKey()).

Bug: Not applicable
Test: Existing tests
Change-Id: I1237b1bc774c64643ece70276301bad53ffeebf2

GitOrigin-RevId: 33b6fc0ab35310804d6ca941c43add6ae4f3c786