[sync] Use fast-path shouldRecommendPluginUpgrade
At the end of the startup activity, or on successful import from the
AndroidModuleDataService, we know that we have models available (and if
we don't, it's unlikely that parsing build files will lead to better
information).  Under those circumstances, we can recommend the plugin
upgrade if necessary without consulting and parsing the
possibly-extensive Gradle build files, which means we no longer need to
execute that check on a background thread.

Bug: n/a
Test: existing, manual.
Change-Id: If7d938540e9298a89f64c3ec89e9546810bf3bef

GitOrigin-RevId: 58ae273000235aea2281b97016121dab4cac5383