Inline the valid use of Device::hasKeyContainedBy
Inline the one valid use of hasKeyContainedBy in AsyncDevicesGetter (as
containsPathOrName) and replace all other uses with
keys.contains(device.getKey()). Delete hasKeyContainedBy.

The one use is valid because the connected devices won't have their
virtual device paths set if the user is using an old emulator. In that
case virtual device names are used to match connected devices with
available virtual devices.

hasKeyContainedBy was also a workaround for nonprefixed keys. As those
don't exist with the switch to PersistentStateComponent it's OK to
delete it.

Bug: Not applicable
Test: Existing tests
Change-Id: I1ae33143170dc01f329e49ceb912734666ca67e2

GitOrigin-RevId: c05f3a0f10984fe59dc9a0987a5c51ebecab6c53