Inline the valid use of Device::matches
Inline the one valid use of matches in AsyncDevicesGetter and replace
all other uses with device.getKey().equals(key). Delete matches.

The one use is valid because the connected devices won't have their
virtual device paths set if the user is using an old emulator. In that
case virtual device names are used to match connected devices with
available virtual devices.

matches was also a workaround for nonprefixed keys. As those don't exist
with the switch to PersistentStateComponent it's OK to delete it.

Bug: Not applicable
Test: Existing tests
Change-Id: I6577576f68d1b243f6a711c2544fed78d714a498