Fix range test for finding possible lambdas
When the first line of a lambda is a comment: the line numbers
from JVMTI will start AFTER the comment, but the bodyExpression
in PSI will include the comment.

Instead of checkint that the body expression is entirely inside
the line interval found with JVMTI, this change will instead
check if the offset range of a lambda is overlapping the offset
range of the found line interval.

Fixes: 177341992
Test: Test adapted to cover this case.
Change-Id: Ied4c9a0b7bd1130b066a9aab37cc7c6a696c9742