Post-install force-stop workaround
This fix works around the issue where if a user has a notification
listener to restart the app after it's stopped, they will see the old
app running instead of sthe updated app.

We fix this issue for installs with --dont-kill, we follow up with
our own manual force-stop, which is synchronous. We enforce that the
the app is not running immediately after install+the added force-stop,
thereby ensuring the subsequent Activity start launches the updated app.

Fixes: 174431980
Test: DeployTaskTest
Change-Id: If7d3b94151d3ed98ac86aef38d7fa83c279dd748

GitOrigin-RevId: c19e6bb3b1b8946ba37f70eec2b103132178d381