Replace the dimension table in layout inspector
This is a workaround for the following problem:
===============================================
JBTable has a new feature for painting the background blue when
hovering over a row. The implementation: TableHoverListener will
simply call repaint on up to 2 rows when the mouse is moved.

BasicTableUI has code to disable repaint of the last row in a
range of rows. The comment for this code is cryptic and I believe
this is actually a bug (in the JDK). This only happens when the
immediate parent of the table is NOT a JScrollPane.

This bug in connection with the new table hover display will cause
rows to "dissapear" i.e. fail to paint when moving the mouse up
over several rows.
===============================================

The fix used in this CL is to make the table a a child of the
existing JScrollPane just like the other tables in the properties
panel. This effectively disables the bug in BasicTableUI.

This change also allows searching in the dimension table.

Fixes: 177357875
Fixes: 177412037
Test: Added test for DimensionBuilder
Change-Id: I83b82728dda8148e9a5d71317bb6e1f698d33f5f