LL: Fix out of order events
1) Ensure we only have one listeners on the editor

The editor listener service gets disabled for various reasons but
the registered listeners persists until the project is disposed.

Therefore, we need to re-enable the listener service on every deploy
but we don't need to register a new listener if there is already one.

Otherwise, we would end up with N events per keystrokes where N
is number of previous deploys.

2) We timestamp each literal updates

In cases where update events is faster than the host can handle them,
it is possible that updates arrived to the installer in wrong order.

To prevent that from happening, we discard literal updates that
is order than the last update we sent to the device.

Bug: 176096647
Test: None
Change-Id: I60a53e96b35430ee1385f4386ff9edf184ab9628

GitOrigin-RevId: c040e8726fcc84fd319a0340ee3e82fdff9da363