Add support for function references
Add resolver logic for function reference locations received from
the agent. It turns out that the synthetic names generated for
the class of these instances is using the same mechanism as lambdas.

As a side note: I made the line numbers passed to LambdaResolver 1
based such that the line numbers are consistent in tests.

Fixes: 177344904
Test: Added tests for resolving source locations
Change-Id: I98487a49c702f48843a0088a8a2e2f1fbfe944b6

GitOrigin-RevId: 0f42fd1b73b5fb448d8b6c50118df307de166eb1