Move getSceneView to design surface
The implementation of getSceneView in the DesignSurface was dropping the
coordinates and ignoring them.
This CL updates the DesignSurface behaviour to the expected one (using x
and y) and updates the Nav Editor interaction to not use x and y since
it did not care about them.

Also add a new method to find the SceneView under the cursor that will
be used in a follow-up CL.

Bug: N/A
Test: Existing test pass
Change-Id: I19289fcb0e5bd0e1ef4ef687d28f6e728bc379fa

GitOrigin-RevId: c9fb78327a8a2bbb8f37f020bfacc600a0582a27