Do not display the boot option
Do not display the boot option in the drop down button if the device is
running nor if it has no snapshots. The boot option indicates how an
available virtual device with snapshots will be booted: Cold Boot, Quick
Boot, or with a snapshot. Once a device is running it doesn't make sense
(currently) to display the boot option in the drop down button because
the device has already been booted.

Devices with no snapshots will always be quick booted so the boot option
is redundant for them

Bug: 120497509
Test: UpdaterTest
Change-Id: Ibd6ee40498d62c59c345317438a5852f49229771