Always expand an invisible root in the component tree
When a tree is using an invisble root the root path is expanded by
the logic in the UI. If the UI is updated i.e updateUI() is called,
there is logic in that accidentally sets the node un expanded in:
BasicTreeUI.updateLayoutCacheExpandedNodesIfNecessary.

When the layout inspector is minimized, the restore action causes
a call of tree.updateUI() which will the hidden root to be set to
unexpanded which results in the tree dissapearing.

Avoid this by explicitly setting the node as expanded.

Fixes: 177870708
Test: Add unit test that calls updateUI() with a hidden root node.
Change-Id: I4145b6469c8d0ab9e2ce14a23a04402c6d91c860