Set application name to default only if it's empty
It was causing application name of companion module module to change.
It shouldn't have been an issue because application name is not used
after this point, and companion module application name is not used
at all.

Unfortunately, it was binded to main module's applicationName
which was binded to main module's packageName (in a tricky way!)
which is actually effectively companion's module package name.

Fixes: 152380894
Test: manual, TemplateTest
Change-Id: I3dc0935188e90992190b3ac0e7d85816dac60562

GitOrigin-RevId: 5766f4e9ebdfc3e24d52b9bb01b3ecd4becd24a3