Update for new Android location scheme
The old class (AndroidLocation) is replaced by
an Abstract version as well as AndroidLocationsSingleton.
This latter class is adequate for Studio.

Update almost all usages except for JPS usage that relies
on jarutils which has a old version of some keystore classes
(moved to sdklib in https://android-review.googlesource.com/c/platform/sdk/+/15281/)
These classes rely on the old AndroidLocation and therefore we cannot
remove this class at the moment.

Also update all usages of AndroidSdkHandler.getInstance which now
requires an instance of AndroidLocationsProvider.

Test: existing tests + updated some
Bug: N/A
Change-Id: Ia808d4b7d011944a0c402916d44c5430d791be7b

GitOrigin-RevId: a7dd1765eab2c038c79025365d29b64d2e1f34b2