[Dsl] Make delete() a part of the GradleDslModel interface
This hoists the functionality up from deleting only individual
properties to allowing deletion of blocks from their model.

As an implementation decision, when running on a GradleFileModel,
delete() deletes the contents of the file but does not remove the file
from the file system.  This is consistent with a view that the Dsl model
allows manipulating the properties, rather than the files that are used
to define them.

Bug: 178682348
Test: added
Change-Id: I4a74724d396fd431c1552f1779ab2d1dd57a1dd2

GitOrigin-RevId: 2cfb2c73a5fc54b44b9f16c19c2f0caf2e1bb8d8