[AUA] Support removal of jni() from android sourceSet
Define removal of a property in a generic way, and then implement
support the removal of jni properties from all AndroidSourceSets in the
build model using that generic infrastructure.

This exercise reveals a couple of ways in which the generic
infrastructure is not yet sufficiently generic.

One is the multiplicity of the Dsl elements being operated on: the
implementation of Move as an operation implied that each property being
moved was a single one, while here the implementation of Remove (since
it is acting on an unbounded set of sourceSets) is operating on a
collection.  That's fine for Remove where the operation is independent
of the element being acted on (the delete() method works for all things
it's called on) but Move/Rename will need either to preserve an order
correspondence between detection and use, or to provide more specific
information to the operator.

Another, related, is the loop over models in the inner
RefactoringProcessor classes of the RefactoringInfo data classes: that
looks as if it should be generic, but the details of what is happening
are at present embedded.  As of this CL we have effectively two
closely-related implementations; a third case should provide sufficient
motivation and evidence for a more generic implementation.

Bug: 178682348
Test: added
Change-Id: Ifb1bc56c413a4bc980f3bb53037e4d01121ffc14