[Dsl] Add getRepresentativeContainedPsiElement to PsiElementHolder
The new method, coming with a default implementation, is intended for
when a PsiElement is needed for a Dsl entity, but that Dsl entity might
be implicit (for example, intermediate Dsl blocks if accessed through
member variables, such as in

  android.sourceSets.main.jniLibs.srcDir 'foo'

where the various blocks such as sourceSets and jniLibs do not have
their own distinct block-like Psi elements).

The only implementation of this interface is in GradleDslBlockModel,
which traverses (breadth-first) the on-file elements of the
corresponding Dsl element in an effort to find a Psi element, and
returns the first that it finds, maximising the chance of finding a
PsiElement as high in the hierarchy as possible.

Bug: 178682348
Test: added
Change-Id: I903d119ce26e7cc99e89be1d33d027c748346a2f

GitOrigin-RevId: 9573ec2102291058ed5670b5a2d3da6269ad0f06