[auto-import] Auto refresh local cache asynchronously
...on a scheduled basis, daily. Also, we update the
in-memory maven class registry subsequently if necessary.

This is to avoid http request when providing fixes to
those unresolved symbols, in read actions.

An application service, `MavenClassRegistryManager`, is
introduced for providing maven class registry, for both
on and off of `auto-import` feature.

The underlying `GMavenIndexRepository` is for providing
the last known maven class registry for instant query.
The freshness of local dish cache and in-memory maven class
registry is guaranteed by auto-refresher, refreshing on
a daily basis.

Also, since we publish a compressed index file to addresss
the download size concern, we read download a gzip
index file now.

Bug: 178716371
     178608328
Test: GMavenIndexRepositoryTest
      GzipUtilsKtTest
Change-Id: I924c3092442933a07c8bc5a00ccbf35283316d08