[AUA] Don't override processor configuration from tool window
Separate out the two distinct parts of initialization of the
Java8SpecialCaseDialog: initializing / updating the state of the
individual component processors appropriately; and providing a
back-action callback for re-creating the dialog from the usage view.
When coming from the tool window, the component processors have been
initialized, at least as far as their enabled state goes, but we still
need to produce the callback.

Fixes: 178484670
Test: manual, added constructor tests
Change-Id: Ia2ba59b4542661e81503bba3c5d3dcfe461cc74e