[PSD] Populate module repositories from project settings too
The new per-Project way of setting up repositories for dependency
resolution involves declarations in the project settings.  A
hyper-correct implementation for the Project Structure Dialog would
inspect the value (implicit or otherwise) of repositoriesMode, and
populate the module repositories differently depending on whether it is
PREFER_SETTINGS or PREFER_PROJECT (or even FAIL_ON_PROJECT_REPOS) but
this superset is a first approximation to the default PREFER_PROJECT
semantics.

Bug: 171990802
Test: manual
Change-Id: I778d183c776c350176d44242c61133e77ebad6c3

GitOrigin-RevId: 5f718e1c4033e0a495358eb29b6343beca5930e3