[Dsl] Don't use getChildren() in delete()
The getChildren method in its current use in getResolvedVariables()
implicitly uses the current state of the Dsl model: the implementation
of getChildren() filters out HIDDEN and TO_BE_REMOVED properties, which
is the right thing to do for questions of interpretation of the Dsl.

However, filtering out those properties is the wrong thing to do when
descending the Dsl tree to implement deletion: it will filter out
exactly those properties that we should be considering.  For existing
uses (where we do not delete whole Dsl blocks at a time) this is not a
problem, as the individual elements will be deleted.  However, when
deleting whole blocks, enabled by a forthcoming change, we must visit
with the removal function all the child elements, which are not
necessarily contained within the PsiElement of the parent block.

Bug: n/a
Test: existing; in subsequent CL
Change-Id: I2d88dbd1ed77cd4aa2343abdfc32a59a37d38186

GitOrigin-RevId: 7a134c451db27a5375b31bf6a84d869e36f57d0d