Show the compose tree as a callstack in the component tree
The component tree should indent only for compose nodes that has
multiple sub nodes. In this way we avoid having the extremely
deep component trees with many nodes having exactly 1 sub node.

We can think of all children of a node as being a callstack.

For now: hide this functionality behind a new flag:
- dynamic.layout.inspector.tree.options

I have not included tests since we may decide to not adopt this
idea if the UX tests show resistance from the users.

Bug: 176980086
Bug: 177468337
Bug: 176980632
Test: TODO
Change-Id: I92e4a620b3159b44203ec9cf994ef267f69548f4