Only update render configuration target when it changes
We were changing the render configuration target on every refresh, passing a
new instance of CompatibilityRenderTarget. This was triggering a resources
modification notification, which would cause the model to be updated. Now,
we're checking the underlying real target before updating the wrapping
CompatibilityRenderTarget, so we can only update if necessary.

This prevents Compose Previews from being inflated multiple times
unnecessarily, which led to the creation of multiple instances of
ComposeViewAdapter and that was causing the Animation Preview not to work
properly.

Fixes: 178707070
Test: Manually verified that a single ComposeViewAdapter is created
Change-Id: Iceea6c67464b835b2b4b1c8acc8852e30fb8fd44

GitOrigin-RevId: 35c61fc41a3f316c6b5dd6dc90e907655c98874e