Fix deploying via bundle for preview devices
When deploying via the bundle, Studio generates a device spec json file
that is then passed via the Android Gradle Plugin to bundle tool, which
then produces the appropriate APKs from the bundle for the device, to
allow Studio to install them.

Before this change, the device spec contains sdk_version, but not the
code name for a preview device. This breaks the APK from-bundle
production when deploying a device with a preview version min-sdk
version to the corresponding preview device, as this makes the device
spec appear as if it were the previous stable API level.

Fixes: 178764656
Test: Expanded unit tests for device spec file generation (AndroidDeviceSpecUtilTest)
Change-Id: I941a101c57863c54da9797e3c9149a3d6858d1f4

GitOrigin-RevId: e31457b08ed5fde8897d4cff23e097e02b432966