Find existing versions from build.gradle repositories
GradleModuleSystem.analyzeDependencyCompatibility used to use the
RepositoryUrlManager to find library versions. While that works
fine for libraries in the GoogleMavenRepository we would not see
libraries from MavenCentral nor from local repositories.

With this change we are using all the repositories found by the
GradleBuildModel which is parsing the build.gradle file.
If the GradleBuildModel wasn't able to parse the repositories we
will fallback to GoogleRepository & MavenCentralRepository.

For the tests we now have to include the pom files for the local
maven repository to be able to identify the versions.

For now: only the depedencies of libraries from GoogleRepository
are analyzed.

Bug: 178244644
Test: Unit test adapted to new implementation
Change-Id: I846bf5cd7087e40e698664ed05df7b31a1fd5afb

GitOrigin-RevId: 0e12f63ca8ea5e2af033193a0942dedf4b390d89