Removes Guava library
This change removes the Guava library from
the tools/adt/idea project. Guava is provided
by the IntelliJ platform instead.

Currently we use a custom Guava as a dependency
but then we don't bundle it and use what is
provided by the platform. In fact both Guava libraries
are in the class path when running tests.

Bug: 179064916
Test: Removing a library, existing tests should pass with the other one
Change-Id: I714c8f26cb85b241abeaea0bf6788cdd4605a71b

GitOrigin-RevId: f52059d1b600d8bccf498db03f87bd4aa11c638e