Move HRenderer implementations to test packages
In an attempt to apply performance optimization to all HRenderer
implementations, I realized they're only used in tests (VisualTest
included) so I'm moving them to the test packages so that
production code doesn't depending on them.

This leaves just CaptureNodeHRenderer (CPU) and HeapSetNodeHRenderer
(native memory), both of which are already optimized.

Bug: n/a
Test: existing
Change-Id: Iba6b3f0aa6fcf589784b5f80b99b7550afce09ab