[UTP AGP] add support for DslName in TestResults
Added support for gradle dsl name in importUtpTestResultAction. For
gradle managed devices the avd name is not the name defined in the dsl
but rather a name that corresponds to the dsl definition so devices can
be shared accross projects. This fix is so that when test results are
imported from a managed device it will show:
Gradle:dslName
instead of:
avdName

Bug:173824067
Test: added test for managed devices to the UtpTestResultAdapter

Change-Id: Ib1220409015704ba0ad4ce34535c81d5d776a924