Do not block LiveLiteralService on waiting for UI
The LiveLiteralsService receives calls from other services from
different threads and then it does actions in the UI thread to
enable/disable highlights. There is no reason for it to wait for those
actions to execute.
Removing the wait prevents dead-locks waiting for the UI thread and
fixes one where the ModuleClassLoaderManager is blocking the UI thread
and waiting in a different thread for it to be unblocked.

Fixes: 179163879
Test: N/A
Change-Id: I3ae1e0abf6edf353e7c6af4970ba9c228b808a85