[auto-import] Prevent unnecessary network requests
...with the help of the cached ETag value.

When requesting gmaven index, we explicitly store the
corresponding ETag values, in a `.properties` file, as
a sibling to the local cached content. So, such cached
Etag value can be an identifier to determine if there's
new changes since the last request, and
`304 Not Modified Response` is the expected response
if we've already gotten an up to date cache.

There's some unfortunate scenarios,
1) Only etag file exists - we still make network requests
as no cached data availble.
2) Cahce data without etag file - we can't judge
if the cache data is up to date, so we still make network
requests.

When no updates are required for the local cache when
refreshing, we log out info: HTTP not modified since the
last request for URL:
https://dl.google.com/android/studio/gmaven/index/release/...
(etag: "843fc7").

Bug: 178608328
Test: GMavenIndexRepositoryImplTest

Change-Id: Id95013466d745f97c776ae5f42ee749aa97042c4

GitOrigin-RevId: ee8ea5bdaeced5b924e9305dd37f9c7c94d033ae