Moves bundled dependencies to plugin def
This change moves the bundled library dependencies
from being defined on each module to the plugin definition.

Additionally this change introduces the same check for libraries
than the one we have for modules: If a module needs a library
dependency, then plugin layout is used to verify that the dependency
is satisfied in production.

Test: Added check at build time
Bug: 179064916
Change-Id: I6eb36894f998327931ca9b6c7963122bddfdccfc

GitOrigin-RevId: 3d2c0be5b235099758251c63a18e8529ad8f9510