Fix AS test compatibility with Gradle Tooling API 6.7+ (part 2)
This change is a follow-up to I94dc9c67de7edee1ab6de5ac1def7bd5be0c6d08,
which brings `studio-plugin-gradle` to the top of module dependency
lists where it has been already present. However, since many tests
that depend on `studio-plugin-Kotlin` also depend on
`studio-plugin-gradle` indirectly the previous change is not enough.

This change adds a compile scope dependency on `studio-plugin-gradle` to
any test module that depends on `studio-plugin-Kotlin` if it also
depends on `studio-plugin-gradle`. The dependency is added before any
direct or transitive dependencies on the Kotlin plugin but preferably
after the dependency on `studio-sdk`.

Any existing test scope dependencies on `studio-plugin-gradle` are
are replaced with compile scope dependencies. This is required to make
sure that no compile scope dependency brings the Kotlin plugin to the
classpath before the Gradle plugin via the production part of an iml
module on which the test part implicitly depends on.

Bug: 179161606
Test: n/a
Change-Id: Icf4ef210bce62aec1e5b5ceafe0a4da63eeac053

GitOrigin-RevId: 4d5a0f5691cbff450aa610bedde258b09845adcb