[Dsl] Replace shouldUseAssignment with syntax
The other half of booleans for using assignment is the information
attached to parsed or newly-created Dsl elements describing whether they
should be using assignment or method-call syntax once written out to a
backend language.  This time, the boolean is two-state, but as we
prepare to support more syntactic forms those two states are no longer
sufficient.  Translate the existing treatment directly into
ExternalNameSyntax enum values: false is METHOD, true is ASSIGNMENT.

Bug: 179483658
Test: existing (no change in functionality)
Change-Id: I145747181965a943b1b56bb8811c205ccc9191c0